Updated readers, server & client

+++ src/fr/umlv/chatos/readers/clientop/ClientMessageOpCode.java
@@ -3,15 +3,15 @@ package fr.umlv.chatos.readers.clientop;
 import java.util.Optional;
 
 public enum ClientMessageOpCode {
-    INITIALIZATION(1),
-    GLOBAL_MESSAGE(4),
-    PERSONAL_MESSAGE(6),
-    PRIVATE_CONNECTION_REQUEST(8),
-    PRIVATE_CONNECTION_RESPONSE(10),
-    PRIVATE_CONNECTION_ESTABLISHMENT(12);
+    INITIALIZATION((byte) 1),
+    GLOBAL_MESSAGE((byte) 4),
+    PERSONAL_MESSAGE((byte) 6),
+    PRIVATE_CONNECTION_REQUEST((byte) 8),
+    PRIVATE_CONNECTION_RESPONSE((byte) 10),
+    PRIVATE_CONNECTION_ESTABLISHMENT((byte) 12);
 
-    private final int value;
-    ClientMessageOpCode(final int value) {
+    private final byte value;
+    ClientMessageOpCode(final byte value) {
         this.value = value;
     }
     public int value() {
@@ -19,7 +19,7 @@ public enum ClientMessageOpCode {
     }
 
 
-    public static Optional<ClientMessageOpCode> clientMessageOpCode(int code) {
+    public static Optional<ClientMessageOpCode> clientMessageOpCode(byte code) {
         for (ClientMessageOpCode opCode : ClientMessageOpCode.values()) {
             if (opCode.value() == code) {
                 return Optional.of(opCode);
