+++ src/fr/umlv/chatos/readers/serverop/ServerErrorCode.java
@@ -3,24 +3,24 @@ package fr.umlv.chatos.readers.serverop;
 import java.util.Optional;
 
 public enum ServerErrorCode {
-    UNDEFINED(0),
-    EMPTY_PSEUDO(1),
-    TOO_LONG_PSEUDO(2),
-    ALREADY_USED(3),
-    NOT_LINKED(4),
-    TOO_LONG_MESSAGE(5),
-    CONNEXION_DECLINED(6),
-    CONNEXION_TIMEOUT(7);
+    UNDEFINED((byte) 0),
+    EMPTY_PSEUDO((byte) 1),
+    TOO_LONG_PSEUDO((byte) 2),
+    ALREADY_USED((byte) 3),
+    NOT_LINKED((byte) 4),
+    TOO_LONG_MESSAGE((byte) 5),
+    CONNEXION_DECLINED((byte) 6),
+    CONNEXION_TIMEOUT((byte) 7);
 
-    private final int value;
-    ServerErrorCode(final int value) {
+    private final byte value;
+    ServerErrorCode(final byte value) {
         this.value = value;
     }
-    public int value() {
+    public byte value() {
         return this.value;
     }
 
-    public static Optional<ServerErrorCode> serverErrorCode(int value){
+    public static Optional<ServerErrorCode> serverErrorCode(byte value){
         for(var opcode : ServerErrorCode.values()){
             if(value == opcode.value) return Optional.of(opcode);
         }
